requirements.txt
google-genai>=0.8.0
pyyaml>=6.0
fastapi>=0.115.0
uvicorn>=0.32.0
python-dotenv>=1.0.0
